refactor: change folder structure for footer components including contacts, feedback form, navigation, and policies
