+++ src/entities/website/Footer/LogoAndSocialsSection/FooterLogoAndSocialsSection.tsx
