+++ src/entities/website/Footer/LogoAndSocialsSection/index.ts
