+++ src/entities/website/Footer/Policy/FooterPolicy.tsx
