+++ src/entities/website/Footer/Policy/index.ts
