+++ src/entities/website/Footer/Policy/services/footerPolicy.service.ts
